fix(consume): keep resolver phase within budget
apm-spec-waiver: Restores existing branch update idempotency without extending OpenAPM behavior.


Copilot-Session: 4dc28902-c462-46b3-b506-32834694e95c
